refactor: extract Universal Cart functionality into dedicated module
- Create project/cart/ module with separate concerns:
  * cart-integration.js: UI handlers and notifications
  * cart-operations.js: Business logic for equipment addition
  * index.js: Module exports and initialization
- Refactor projects-view.js to use new module (reduced from 439 to 57 lines)
- Eliminate code duplication by extracting ~380 lines into reusable modules
- Maintain backward compatibility with global window functions
- Improve error handling and batch operations
- Add comprehensive documentation and README
- Fix Universal Cart import path issue
